# docs/openapi-public.yaml
...
openapi: 3.0.3
info:
  title: ChangeNOW Dashboard API - public endpoints
  version: 7.2.3
  description: Generated from NestJS controllers. Includes only the latest version
    of each endpoint that is not protected by ApiKey*Guard. Endpoints with AuthGuard/JWT
    guards are marked with bearerAuth/cookieAuth security.
servers:
- url: "/"
  description: Relative base URL. Nest URI versioning is represented as /v{version}/...
    in paths.
...
paths:
...
  "/v1.0/cashback/estimate":
    get:
      tags:
      - cashback
      operationId: cashback_get_v1_0_cashback_estimate_estimate
      summary: estimate
      description: Public API endpoint not protected by ApiKey*Guard. Generated from
        NestJS controller metadata.
      x-source:
        file: apps/cashback/src/cashback-api/cashback.controller.ts
        line: 22
        guards:
        - JwtAuthGuard
        - Jwt2faAuthGuard
        - JwtCookieAuthGuard
        returnType: ''
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      - cookieAuth: []
...
  "/v2.0/exchange/estimate":
    get:
      tags:
      - tx-estimation
      operationId: tx_estimation_get_v2_0_exchange_estimate_getEstimateV2
      summary: getEstimateV2
      description: Public API endpoint not protected by ApiKey*Guard. Generated from
        NestJS controller metadata.
      x-source:
        file: apps/tx-estimation/src/tx-estimation.controller.ts
        line: 95
        guards:
        - JwtAuthGuard
        - Jwt2faAuthGuard
        - JwtCookieAuthGuard
        returnType: ''
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
      security:
      - bearerAuth: []
      - cookieAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    cookieAuth:
      type: apiKey
      in: cookie
      name: cn_at
      x-legacy-cookie-name: accessToken
    csrfHeader:
      type: apiKey
      in: header
      name: x-csrf-token
      description: Required only for operations with CsrfGuard. The value must match
        the cn_csrf cookie.
...